[Fix] `membership_degrees_hist` is now correct: it was shifted (same with `clusters_parameters_hist`)
The shift was caused by `slc` slice variable. For `hist_output` it should be `slice(-n_test_dates - 1, -1)`, so that the forecast for the future date (which is not known yet) is not included into the `hist_output` series. But `membership_degrees_hist` shows data not for future dates, but for current dates, thus it needs `slice(-n_test_dates, None)`