[Feature] Implemented consequent estimation on `n_points_fitting` last points
It is important to notice that there must have been implemented some logic on `first_h` calculation at each estimation step. For this purpose `first_h_current` attribute was introduced, and after each push (`_push()`) it is set as first `starting_index` (e.g., 1) elements of sliced `train_data`. The slice is `[-n_points_fitting:]`. `first_h_current` at the initial fitting can be set to a custom value, but it must be passed explicitly to the constructor as the `first_h` parameter.

Also, uncommented usage of estimated at the previous step consequent parameters as initial value for the next step estimation